requirements.txt
arrow==0.7.0
beautifulsoup4==4.4.1
disqus==0.0.4
Django==1.10.2
django-allauth==0.25.2
django-allauth-bootstrap==0.1
django-emoticons==1.1.2
django-forms-bootstrap==3.0.1
django-tinymce==2.3.0
evernote==1.25.1
httplib2==0.9.2
oauth2==1.9.0.post1
oauthlib==1.1.1
Pillow==3.2.0
python-dateutil==2.5.3
python-openid==2.2.5
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
